third_party/blink/web_tests/external/wpt/webcodecs/video-encoder-rescaling.https.any.worker_vp9_p0-expected.txt

This is a testharness.js-based test.
[FAIL] Scaling Image in Encoding from 64 x 64 to 128 x 128 Format: RGBX
  assert_approx_equals: Scaled Image differs too much! Scaling from 64 x 64 to 128 x 128 Format:RGBX expected 0 +/- 0.05 but got NaN
[FAIL] Scaling Image in Encoding from 64 x 64 to 128 x 128 Format: I420
  assert_approx_equals: Scaled Image differs too much! Scaling from 64 x 64 to 128 x 128 Format:I420 expected 0 +/- 0.05 but got NaN
[FAIL] Scaling Image in Encoding from 128 x 128 to 128 x 128 Format: RGBX
  assert_approx_equals: Scaled Image differs too much! Scaling from 128 x 128 to 128 x 128 Format:RGBX expected 0 +/- 0.05 but got NaN
[FAIL] Scaling Image in Encoding from 128 x 128 to 128 x 128 Format: I420
  assert_approx_equals: Scaled Image differs too much! Scaling from 128 x 128 to 128 x 128 Format:I420 expected 0 +/- 0.05 but got NaN
[FAIL] Scaling Image in Encoding from 128 x 128 to 64 x 64 Format: RGBX
  assert_approx_equals: Scaled Image differs too much! Scaling from 128 x 128 to 64 x 64 Format:RGBX expected 0 +/- 0.05 but got NaN
[FAIL] Scaling Image in Encoding from 128 x 128 to 64 x 64 Format: I420
  assert_approx_equals: Scaled Image differs too much! Scaling from 128 x 128 to 64 x 64 Format:I420 expected 0 +/- 0.05 but got NaN
[FAIL] Scaling Image in Encoding from 32 x 32 to 96 x 96 Format: RGBX
  assert_approx_equals: Scaled Image differs too much! Scaling from 32 x 32 to 96 x 96 Format:RGBX expected 0 +/- 0.05 but got NaN
[FAIL] Scaling Image in Encoding from 32 x 32 to 96 x 96 Format: I420
  assert_approx_equals: Scaled Image differs too much! Scaling from 32 x 32 to 96 x 96 Format:I420 expected 0 +/- 0.05 but got NaN
[FAIL] Scaling Image in Encoding from 192 x 192 to 64 x 64 Format: RGBX
  assert_approx_equals: Scaled Image differs too much! Scaling from 192 x 192 to 64 x 64 Format:RGBX expected 0 +/- 0.05 but got NaN
[FAIL] Scaling Image in Encoding from 192 x 192 to 64 x 64 Format: I420
  assert_approx_equals: Scaled Image differs too much! Scaling from 192 x 192 to 64 x 64 Format:I420 expected 0 +/- 0.05 but got NaN
[FAIL] Scaling Image in Encoding from 64 x 32 to 128 x 64 Format: RGBX
  assert_approx_equals: Scaled Image differs too much! Scaling from 64 x 32 to 128 x 64 Format:RGBX expected 0 +/- 0.05 but got NaN
[FAIL] Scaling Image in Encoding from 64 x 32 to 128 x 64 Format: I420
  assert_approx_equals: Scaled Image differs too much! Scaling from 64 x 32 to 128 x 64 Format:I420 expected 0 +/- 0.05 but got NaN
[FAIL] Scaling Image in Encoding from 128 x 256 to 64 x 128 Format: RGBX
  assert_approx_equals: Scaled Image differs too much! Scaling from 128 x 256 to 64 x 128 Format:RGBX expected 0 +/- 0.05 but got NaN
[FAIL] Scaling Image in Encoding from 128 x 256 to 64 x 128 Format: I420
  assert_approx_equals: Scaled Image differs too much! Scaling from 128 x 256 to 64 x 128 Format:I420 expected 0 +/- 0.05 but got NaN
[FAIL] Scaling Image in Encoding from 64 x 64 to 128 x 192 Format: RGBX
  assert_approx_equals: Scaled Image differs too much! Scaling from 64 x 64 to 128 x 192 Format:RGBX expected 0 +/- 0.05 but got NaN
[FAIL] Scaling Image in Encoding from 64 x 64 to 128 x 192 Format: I420
  assert_approx_equals: Scaled Image differs too much! Scaling from 64 x 64 to 128 x 192 Format:I420 expected 0 +/- 0.05 but got NaN
[FAIL] Scaling Image in Encoding from 128 x 192 to 64 x 64 Format: RGBX
  assert_approx_equals: Scaled Image differs too much! Scaling from 128 x 192 to 64 x 64 Format:RGBX expected 0 +/- 0.05 but got NaN
[FAIL] Scaling Image in Encoding from 128 x 192 to 64 x 64 Format: I420
  assert_approx_equals: Scaled Image differs too much! Scaling from 128 x 192 to 64 x 64 Format:I420 expected 0 +/- 0.05 but got NaN
Harness: the test ran to completion.
